+++ esp_cam_sensor/include/esp_cam_sensor_types.h
@@ -74,6 +74,7 @@ typedef enum {
 
 #define ESP_CAM_SENSOR_STATS_FLAG_WB_GAIN           (1 <<  0)
 #define ESP_CAM_SENSOR_STATS_FLAG_AGC_GAIN          (1 <<  1)
+#define ESP_CAM_SENSOR_STATS_FLAG_EXPOSURE          (1 <<  2)
 
 #define ESP_CAM_SENSOR_PARAM_TYPE_NUMBER            1
 #define ESP_CAM_SENSOR_PARAM_TYPE_BITMASK           2
@@ -437,7 +438,8 @@ typedef struct {
 typedef struct {
     uint32_t flags;
     uint32_t seq;
-    uint16_t agc_gain; /*!< AGC gain output to sensor */
+    uint32_t aec_exp;  /*!< AEC exposure output to sensor */
+    volatile float agc_gain; /*!< AGC gain output to sensor */
     union {
         struct {
             uint8_t red_avg;