+++ esp_video/CHANGELOG.md
@@ -1,7 +1,9 @@
 ## Unreleased
 
+- Add RISC-V software swap byte function for DVP video device
 - Add the customized video class to call esp_cam_sensor ioctl commands directly
 - Add the second SPI video device
+- Add camera sensor exposure parameter to statistics
 - Improve AEC controller time precision
 - Fix the CSI video device initializes twice when starting up
 - Update to use esp_cam_sensor v1.4.x